requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
openai>=1.3.0
python-dotenv>=1.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.22.0
fastapi>=0.100.0
uvicorn>=0.23.0
flask-cors>=3.0.10
streamlit>=1.22.0
python-multipart>=0.0.5
langchain>=0.0.267
# Video generation dependencies
moviepy>=1.0.3
gTTS>=2.3.1
